Update the group and package
