+++ kopykat-ksp/src/main/kotlin/at/kopyk/utils/lang/JoinWithWhen.kt
@@ -1,0 +1,4 @@
+package at.kopyk.utils.lang
+
+internal fun <A> Sequence<A>.joinWithWhen(subject: String = "this", transform: (A) -> String = { it.toString() }) =
+  joinToString(prefix = "when ($subject) {\n", separator = "\n", postfix = "\n}", transform = transform)
