--- ksp/src/main/kotlin/fp/serrano/kopykat/KopyKatProvider.kt
@@ -1,13 +1,0 @@
-@file:Suppress("WildcardImport")
-package fp.serrano.kopykat
-
-import com.google.devtools.ksp.processing.*
-
-public class KopyKatProvider : SymbolProcessorProvider {
-  override fun create(environment: SymbolProcessorEnvironment): SymbolProcessor =
-    KopyKatProcessor(
-      codegen = environment.codeGenerator,
-      logger = environment.logger,
-      options = KopyKatOptions.fromKspOptions(environment.options)
-    )
-}
