--- ksp/src/main/kotlin/fp/serrano/kopykat/utils/lang/JoinWithWhen.kt
@@ -1,4 +1,0 @@
-package fp.serrano.kopykat.utils.lang
-
-internal fun <A> Sequence<A>.joinWithWhen(subject: String = "this", transform: (A) -> String = { it.toString() }) =
-  joinToString(prefix = "when ($subject) {\n", separator = "\n", postfix = "\n}", transform = transform)
