--- ksp/src/test/kotlin/fp/serrano/kopykat/Compiler.kt
@@ -1,9 +1,0 @@
-package fp.serrano.kopykat
-
-internal fun String.failsWith(check: (String) -> Boolean) {
-  failsWith(KopyKatProvider(), check)
-}
-
-internal fun String.evals(vararg things: Pair<String, Any?>) {
-  evals(KopyKatProvider(), *things)
-}
